Extract timeline filters and harden timeline ID parsing
