+++ frontend/src/pages/RecipientRecord/pages/Timeline.tsx
@@ -3,20 +3,19 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Alert, Checkbox, Dropdown } from '@trussworks/react-uswds';
 import React, { useMemo, useRef, useState } from 'react';
 import { Helmet } from 'react-helmet';
-import { v4 as uuidv4 } from 'uuid';
-import { DATE_CONDITIONS, EMPTY_MULTI_SELECT, FILTER_CONDITIONS } from '../../../Constants';
 import Container from '../../../components/Container';
 import Drawer from '../../../components/Drawer';
 import DrawerTriggerButton from '../../../components/DrawerTriggerButton';
-import { purposeFilter } from '../../../components/filter/communicationLogFilters';
-import FilterDateRange from '../../../components/filter/FilterDateRange';
 import FilterPanel from '../../../components/filter/FilterPanel';
-import FilterSelect from '../../../components/filter/FilterSelect';
-import { goalCategoryFilter } from '../../../components/filter/goalFilters';
+import {
+  createInitialTimelineFilters,
+  serializeTimelineFilter,
+  TIMELINE_FILTER_CONFIG,
+  type TimelineFilter,
+} from '../../../components/filter/timelineFilters';
 import NoResultsFound from '../../../components/NoResultsFound';
 import { getRecipientTimeline } from '../../../fetchers/recipient';
 import useFetch from '../../../hooks/useFetch';
-import { formatDateRange } from '../../../utils';
 import './Timeline.css';
 
 interface TimelineProps {
@@ -24,133 +23,18 @@ interface TimelineProps {
   regionId: string;
 }
 
-interface TimelineFilter {
-  id: string;
-  topic: string;
-  condition: string;
-  query: string | string[];
-}
-
 interface TimelineResponse {
   count: number;
   events: unknown[];
 }
 
-const LAST_TWELVE_MONTHS = formatDateRange({
-  lastTwelveMonths: true,
-  forDateTime: true,
-});
-
-const TIMELINE_DATE_OPTIONS = [
-  {
-    label: 'Last three months',
-    value: formatDateRange({ lastThreeMonths: true, forDateTime: true }),
-  },
-  {
-    label: 'Last six months',
-    value: formatDateRange({ lastSixMonths: true, forDateTime: true }),
-  },
-  {
-    label: 'Last twelve months',
-    value: LAST_TWELVE_MONTHS,
-  },
-];
-
-const EVENT_TYPES = [
-  'Email communication',
-  'Phone communication',
-  'In person communication',
-  'Virtual communication',
-  'TTA activity',
-  'Training session',
-  'Goal added',
-  'Goal suspended',
-  'Goal closed',
-  'Goal reopened',
-  'TTA request',
-  'Monitoring report received',
-].map((label) => ({ label, value: label }));
-
-const dateFilter = {
-  id: 'date',
-  display: 'Date',
-  conditions: DATE_CONDITIONS,
-  defaultValues: {
-    is: LAST_TWELVE_MONTHS,
-    'is within': '',
-    'is on or after': '',
-    'is on or before': '',
-  },
-  displayQuery: (query: string | string[]) => {
-    const value = Array.isArray(query) ? query.join(', ') : query;
-    if (value.includes('-')) {
-      return formatDateRange({ string: value, withSpaces: true });
-    }
-    return value;
-  },
-  renderInput: (
-    _id: string,
-    condition: string,
-    query: string | string[],
-    onApplyQuery: (value: string) => void
-  ) => (
-    <FilterDateRange
-      condition={condition}
-      query={query}
-      onApplyDateRange={onApplyQuery}
-      customDateOptions={TIMELINE_DATE_OPTIONS}
-    />
-  ),
-};
-
-const eventTypeFilter = {
-  id: 'eventType',
-  display: 'TTA event type',
-  conditions: FILTER_CONDITIONS,
-  defaultValues: EMPTY_MULTI_SELECT,
-  displayQuery: (query: string | string[]) => [query].flat().join(', '),
-  renderInput: (
-    id: string,
-    condition: string,
-    query: string[],
-    onApplyQuery: (value: string[]) => void
-  ) => (
-    <FilterSelect
-      onApply={onApplyQuery}
-      inputId={`event-type-${condition}-${id}`}
-      labelText="Select TTA event types to filter by"
-      options={EVENT_TYPES}
-      selectedValues={query}
-    />
-  ),
-};
-
-const TIMELINE_FILTER_CONFIG = [
-  dateFilter,
-  { ...purposeFilter, display: 'Communication purpose' },
-  goalCategoryFilter,
-  eventTypeFilter,
-].sort((a, b) => a.display.localeCompare(b.display));
-
-const INITIAL_FILTERS: TimelineFilter[] = [
-  {
-    id: uuidv4(),
-    topic: 'date',
-    condition: 'is within',
-    query: LAST_TWELVE_MONTHS,
-  },
-];
-
-const serializeFilter = ({ topic, condition, query }: TimelineFilter) =>
-  JSON.stringify({ topic, condition, query });
-
 export default function Timeline({ recipientId, regionId }: TimelineProps): React.ReactElement {
   const aboutDrawerRef = useRef<HTMLButtonElement>(null);
-  const [filters, setFilters] = useState<TimelineFilter[]>(INITIAL_FILTERS);
+  const [filters, setFilters] = useState<TimelineFilter[]>(createInitialTimelineFilters);
   const [direction, setDirection] = useState<'asc' | 'desc'>('desc');
   const [hideMultiRecipientCommunications, setHideMultiRecipientCommunications] = useState(false);
 
-  const serializedFilters = useMemo(() => filters.map(serializeFilter), [filters]);
+  const serializedFilters = useMemo(() => filters.map(serializeTimelineFilter), [filters]);
 
   const { data, error, loading } = useFetch(
     { count: 0, events: [] } as TimelineResponse,
